apiCreation/python/django/challenge/requirements.txt
Django==2.0.6
django-compressor==2.2
djangorestframework==3.8.2
pymongo==3.6.1
psycopg2==2.7.5
